pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyraml"
version = "0.1.0"
description = "A deep learning framework focusing on simplicity and educational purposes"
authors = [{ name = "Your Name" }]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "numba>=0.56.0",
    "llvmlite>=0.39.0",
    "cupy-cuda11x>=11.0.0",
    "ray>=2.0.0",
    "networkx>=2.7.0",
    "opt_einsum>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "mypy>=0.950",
    "sphinx>=4.5.0",
]
cuda = [
    "cupy-cuda11x>=11.0.0",
    "cuda-python>=11.7",
]
